backend/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
pillow
numpy
pandas
httpx
python-dotenv
